+++ Engine/Source/Runtime/Core/Private/GenericPlatform/GenericPlatformMemory.cpp
@@ -258,13 +258,13 @@ void FGenericPlatformMemory::DumpStats( class FOutputDevice& Ar )
 	const FName CategoryName(TEXT("LogMemory"));
 #endif
 	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Platform Memory Stats for %s"), ANSI_TO_TCHAR(FPlatformProperties::PlatformName()));
-	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Process Physical Memory: %.2f MB used, %.2f MB peak"), MemoryStats.UsedPhysical*InvMB, MemoryStats.PeakUsedPhysical*InvMB);
-	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Process Virtual Memory: %.2f MB used, %.2f MB peak"), MemoryStats.UsedVirtual*InvMB, MemoryStats.PeakUsedVirtual*InvMB);
+	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Process Physical Memory: %.2f MB used, %.2f MB peak"), (float)MemoryStats.UsedPhysical*InvMB, (float)MemoryStats.PeakUsedPhysical*InvMB);
+	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Process Virtual Memory: %.2f MB used, %.2f MB peak"), (float)MemoryStats.UsedVirtual*InvMB, (float)MemoryStats.PeakUsedVirtual*InvMB);
 
 	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Physical Memory: %.2f MB used,  %.2f MB free, %.2f MB total"), 
-		(MemoryStats.TotalPhysical - MemoryStats.AvailablePhysical)*InvMB, MemoryStats.AvailablePhysical*InvMB, MemoryStats.TotalPhysical*InvMB);
+		(float)(MemoryStats.TotalPhysical - MemoryStats.AvailablePhysical)*InvMB, (float)MemoryStats.AvailablePhysical*InvMB, (float)MemoryStats.TotalPhysical*InvMB);
 	Ar.CategorizedLogf(CategoryName, ELogVerbosity::Log, TEXT("Virtual Memory: %.2f MB used,  %.2f MB free, %.2f MB total"), 
-		(MemoryStats.TotalVirtual - MemoryStats.AvailableVirtual)*InvMB, MemoryStats.AvailableVirtual*InvMB, MemoryStats.TotalVirtual*InvMB);
+		(float)(MemoryStats.TotalVirtual - MemoryStats.AvailableVirtual)*InvMB, (float)MemoryStats.AvailableVirtual*InvMB, (float)MemoryStats.TotalVirtual*InvMB);
 	}
 }
